style(work-localisation): design localisation
