+++ yaki_mobile/assets/translations/fr.json
@@ -100,7 +100,7 @@
   "morning": "Matin",
   "afternoon": "Après-midi",
   "IWorkthisMorning": "Je travaille ce",
-  "IWorkthisAfternoon": "Je travaille cette",
+  "IWorkthisAfternoon": "Je travaille cet",
   "Iam": "Je suis",
   "office": "Au Bureau",
   "remote": "En Télétravail",
